Split ATX handler to prepare for handling ATX v2
