+++ activation/handler_test.go
@@ -114,8 +114,8 @@ func toAtx(t testing.TB, watx *wire.ActivationTxV1) *types.ActivationTx {
 	return atx
 }
 
-type testHandler struct {
-	*Handler
+type handlerMocks struct {
+	goldenATXID types.ATXID
 
 	mclock     *MocklayerClock
 	mpub       *pubsubmocks.MockPublisher
@@ -125,13 +125,22 @@ type testHandler struct {
 	mtortoise  *mocks.MockTortoise
 }
 
+type testHandler struct {
+	*Handler
+
+	cdb        *datastore.CachedDB
+	edVerifier *signing.EdVerifier
+
+	handlerMocks
+}
+
 type atxHandleOpts struct {
 	postVerificationDuration time.Duration
 	poetLeaves               uint64
 	distributedPost          bool
 }
 
-func (h *testHandler) expectAtxV1(atx *wire.ActivationTxV1, nodeId types.NodeID, opts ...func(*atxHandleOpts)) {
+func (h *handlerMocks) expectAtxV1(atx *wire.ActivationTxV1, nodeId types.NodeID, opts ...func(*atxHandleOpts)) {
 	settings := atxHandleOpts{}
 	for _, opt := range opts {
 		opt(&settings)
@@ -167,657 +176,57 @@ func (h *testHandler) expectAtxV1(atx *wire.ActivationTxV1, nodeId types.NodeID,
 	h.mtortoise.EXPECT().OnAtx(gomock.Any(), gomock.Any(), gomock.Any())
 }
 
-func newTestHandler(tb testing.TB, goldenATXID types.ATXID) *testHandler {
+type testHandlerOptions struct {
+	tickSize uint64
+}
+
+func newTestHandlerMocks(tb testing.TB, golden types.ATXID) handlerMocks {
+	ctrl := gomock.NewController(tb)
+	return handlerMocks{
+		goldenATXID: golden,
+		mclock:      NewMocklayerClock(ctrl),
+		mpub:        pubsubmocks.NewMockPublisher(ctrl),
+		mockFetch:   mocks.NewMockFetcher(ctrl),
+		mValidator:  NewMocknipostValidator(ctrl),
+		mbeacon:     NewMockAtxReceiver(ctrl),
+		mtortoise:   mocks.NewMockTortoise(ctrl),
+	}
+}
+
+func newTestHandler(tb testing.TB, goldenATXID types.ATXID, opts ...func(*testHandlerOptions)) *testHandler {
 	lg := logtest.New(tb)
 	cdb := datastore.NewCachedDB(sql.InMemory(), lg)
-	ctrl := gomock.NewController(tb)
-	mclock := NewMocklayerClock(ctrl)
-	mpub := pubsubmocks.NewMockPublisher(ctrl)
-	mockFetch := mocks.NewMockFetcher(ctrl)
-	mValidator := NewMocknipostValidator(ctrl)
+	edVerifier := signing.NewEdVerifier()
 
-	mbeacon := NewMockAtxReceiver(ctrl)
-	mtortoise := mocks.NewMockTortoise(ctrl)
-
+	options := testHandlerOptions{
+		tickSize: 1,
+	}
+	for _, opt := range opts {
+		opt(&options)
+	}
+	mocks := newTestHandlerMocks(tb, goldenATXID)
 	atxHdlr := NewHandler(
 		"localID",
 		cdb,
 		atxsdata.New(),
-		signing.NewEdVerifier(),
-		mclock,
-		mpub,
-		mockFetch,
-		1,
+		edVerifier,
+		mocks.mclock,
+		mocks.mpub,
+		mocks.mockFetch,
+		options.tickSize,
 		goldenATXID,
-		mValidator,
-		mbeacon,
-		mtortoise,
+		mocks.mValidator,
+		mocks.mbeacon,
+		mocks.mtortoise,
 		lg,
 	)
 	return &testHandler{
-		Handler: atxHdlr,
+		Handler:    atxHdlr,
+		cdb:        cdb,
+		edVerifier: edVerifier,
 
-		mclock:     mclock,
-		mpub:       mpub,
-		mockFetch:  mockFetch,
-		mValidator: mValidator,
-		mbeacon:    mbeacon,
-		mtortoise:  mtortoise,
+		handlerMocks: mocks,
 	}
-}
-
-func TestHandler_SyntacticallyValidateAtx(t *testing.T) {
-	sig, err := signing.NewEdSigner()
-	require.NoError(t, err)
-
-	goldenATXID := types.RandomATXID()
-
-	setup := func(t *testing.T) (hdlr *testHandler, prev, pos *wire.ActivationTxV1) {
-		atxHdlr := newTestHandler(t, goldenATXID)
-
-		prevAtx := newInitialATXv1(t, goldenATXID)
-		prevAtx.NumUnits = 100
-		prevAtx.Sign(sig)
-		atxHdlr.expectAtxV1(prevAtx, sig.NodeID())
-		require.NoError(t, atxHdlr.HandleGossipAtx(context.Background(), "", codec.MustEncode(prevAtx)))
-
-		otherSig, err := signing.NewEdSigner()
-		require.NoError(t, err)
-
-		posAtx := newInitialATXv1(t, goldenATXID)
-		posAtx.Sign(otherSig)
-		atxHdlr.expectAtxV1(posAtx, otherSig.NodeID())
-		require.NoError(t, atxHdlr.HandleGossipAtx(context.Background(), "", codec.MustEncode(posAtx)))
-		return atxHdlr, prevAtx, posAtx
-	}
-	t.Run("valid atx", func(t *testing.T) {
-		t.Parallel()
-		atxHdlr, prevAtx, posAtx := setup(t)
-
-		atx := newChainedActivationTxV1(t, atxHdlr.goldenATXID, prevAtx, posAtx.ID())
-		atx.PositioningATXID = posAtx.ID()
-		atx.Sign(sig)
-
-		atxHdlr.mclock.EXPECT().CurrentLayer().Return(atx.PublishEpoch.FirstLayer())
-		require.NoError(t, atxHdlr.syntacticallyValidate(context.Background(), atx))
-
-		atxHdlr.mValidator.EXPECT().
-			NIPost(gomock.Any(), atx.SmesherID, goldenATXID, gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).
-			Return(1234, nil)
-		atxHdlr.mValidator.EXPECT().NIPostChallengeV1(gomock.Any(), gomock.Any(), atx.SmesherID)
-		atxHdlr.mValidator.EXPECT().PositioningAtx(atx.PositioningATXID, gomock.Any(), goldenATXID, gomock.Any())
-		leaves, units, proof, err := atxHdlr.syntacticallyValidateDeps(context.Background(), atx)
-		require.NoError(t, err)
-		require.Equal(t, uint64(1234), leaves)
-		require.Equal(t, atx.NumUnits, units)
-		require.Nil(t, proof)
-	})
-	t.Run("valid atx with new VRF nonce", func(t *testing.T) {
-		t.Parallel()
-		atxHdlr, prevAtx, posAtx := setup(t)
-
-		newNonce := *prevAtx.VRFNonce + 100
-		atx := newChainedActivationTxV1(t, goldenATXID, prevAtx, posAtx.ID())
-		atx.VRFNonce = &newNonce
-		atx.Sign(sig)
-
-		atxHdlr.mclock.EXPECT().CurrentLayer().Return(atx.PublishEpoch.FirstLayer())
-		require.NoError(t, atxHdlr.syntacticallyValidate(context.Background(), atx))
-
-		atxHdlr.mValidator.EXPECT().
-			NIPost(gomock.Any(), gomock.Any(), goldenATXID, gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).
-			Return(1234, nil)
-		atxHdlr.mValidator.EXPECT().NIPostChallengeV1(gomock.Any(), gomock.Any(), gomock.Any())
-		atxHdlr.mValidator.EXPECT().PositioningAtx(atx.PositioningATXID, gomock.Any(), goldenATXID, atx.PublishEpoch)
-		atxHdlr.mValidator.EXPECT().
-			VRFNonce(gomock.Any(), goldenATXID, newNonce, gomock.Any(), atx.NumUnits)
-		leaves, units, proof, err := atxHdlr.syntacticallyValidateDeps(context.Background(), atx)
-		require.NoError(t, err)
-		require.Equal(t, uint64(1234), leaves)
-		require.Equal(t, atx.NumUnits, units)
-		require.Nil(t, proof)
-	})
-	t.Run("valid atx with decreasing num units", func(t *testing.T) {
-		t.Parallel()
-		atxHdlr, prevAtx, posAtx := setup(t)
-
-		atx := newChainedActivationTxV1(t, goldenATXID, prevAtx, posAtx.ID())
-		atx.NumUnits = prevAtx.NumUnits - 10
-		atx.Sign(sig)
-
-		atxHdlr.mclock.EXPECT().CurrentLayer().Return(atx.PublishEpoch.FirstLayer())
-		require.NoError(t, atxHdlr.syntacticallyValidate(context.Background(), atx))
-		atxHdlr.mValidator.EXPECT().
-			NIPost(gomock.Any(), gomock.Any(), goldenATXID, gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).
-			Return(uint64(1234), nil)
-		atxHdlr.mValidator.EXPECT().NIPostChallengeV1(gomock.Any(), gomock.Any(), gomock.Any())
-		atxHdlr.mValidator.EXPECT().PositioningAtx(atx.PositioningATXID, gomock.Any(), goldenATXID, atx.PublishEpoch)
-		leaves, units, proof, err := atxHdlr.syntacticallyValidateDeps(context.Background(), atx)
-		require.NoError(t, err)
-		require.Equal(t, uint64(1234), leaves)
-		require.Equal(t, atx.NumUnits, units)
-		require.Nil(t, proof)
-	})
-	t.Run("atx with increasing num units, no new VRF, old valid", func(t *testing.T) {
-		t.Parallel()
-		atxHdlr, prevAtx, posAtx := setup(t)
-
-		atx := newChainedActivationTxV1(t, goldenATXID, prevAtx, posAtx.ID())
-		atx.NumUnits = prevAtx.NumUnits + 10
-		atx.Sign(sig)
-
-		atxHdlr.mclock.EXPECT().CurrentLayer().Return(atx.PublishEpoch.FirstLayer())
-		require.NoError(t, atxHdlr.syntacticallyValidate(context.Background(), atx))
-		atxHdlr.mValidator.EXPECT().
-			NIPost(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).
-			Return(uint64(1234), nil)
-		atxHdlr.mValidator.EXPECT().NIPostChallengeV1(gomock.Any(), gomock.Any(), gomock.Any())
-		atxHdlr.mValidator.EXPECT().PositioningAtx(atx.PositioningATXID, gomock.Any(), gomock.Any(), gomock.Any())
-		atxHdlr.mValidator.EXPECT().VRFNonce(gomock.Any(), goldenATXID, *prevAtx.VRFNonce, gomock.Any(), atx.NumUnits)
-		leaves, units, proof, err := atxHdlr.syntacticallyValidateDeps(context.Background(), atx)
-		require.NoError(t, err)
-		require.Equal(t, uint64(1234), leaves)
-		require.Equal(t, prevAtx.NumUnits, units)
-		require.Nil(t, proof)
-	})
-	t.Run("atx with increasing num units, no new VRF, old invalid for new size", func(t *testing.T) {
-		t.Parallel()
-		atxHdlr, prevAtx, posAtx := setup(t)
-
-		atx := newChainedActivationTxV1(t, goldenATXID, prevAtx, posAtx.ID())
-		atx.NumUnits = prevAtx.NumUnits + 10
-		atx.Sign(sig)
-
-		atxHdlr.mclock.EXPECT().CurrentLayer().Return(atx.PublishEpoch.FirstLayer())
-		require.NoError(t, atxHdlr.syntacticallyValidate(context.Background(), atx))
-
-		atxHdlr.mValidator.EXPECT().NIPostChallengeV1(gomock.Any(), gomock.Any(), gomock.Any())
-		atxHdlr.mValidator.EXPECT().
-			VRFNonce(gomock.Any(), goldenATXID, *prevAtx.VRFNonce, gomock.Any(), atx.NumUnits).
-			Return(errors.New("invalid VRF"))
-		_, _, proof, err := atxHdlr.syntacticallyValidateDeps(context.Background(), atx)
-		require.ErrorContains(t, err, "invalid VRF")
-		require.Nil(t, proof)
-	})
-	t.Run("valid initial atx", func(t *testing.T) {
-		t.Parallel()
-		atxHdlr, _, posAtx := setup(t)
-
-		ctxID := posAtx.ID()
-		atx := newInitialATXv1(t, goldenATXID)
-		atx.CommitmentATXID = &ctxID
-		atx.Sign(sig)
-
-		atxHdlr.mclock.EXPECT().CurrentLayer().Return(atx.PublishEpoch.FirstLayer())
-		atxHdlr.mValidator.EXPECT().
-			Post(gomock.Any(), gomock.Any(), ctxID, gomock.Any(), gomock.Any(), atx.NumUnits, gomock.Any())
-		atxHdlr.mValidator.EXPECT().VRFNonce(sig.NodeID(), ctxID, *atx.VRFNonce, gomock.Any(), atx.NumUnits)
-		require.NoError(t, atxHdlr.syntacticallyValidate(context.Background(), atx))
-
-		atxHdlr.mValidator.EXPECT().InitialNIPostChallengeV1(gomock.Any(), gomock.Any(), goldenATXID)
-		atxHdlr.mValidator.EXPECT().
-			NIPost(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).
-			Return(uint64(777), nil)
-		atxHdlr.mValidator.EXPECT().PositioningAtx(atx.PositioningATXID, gomock.Any(), goldenATXID, atx.PublishEpoch)
-		leaves, units, proof, err := atxHdlr.syntacticallyValidateDeps(context.Background(), atx)
-		require.NoError(t, err)
-		require.Equal(t, uint64(777), leaves)
-		require.Equal(t, atx.NumUnits, units)
-		require.Nil(t, proof)
-	})
-	t.Run("atx targeting wrong publish epoch", func(t *testing.T) {
-		t.Parallel()
-		atxHdlr, prevAtx, posAtx := setup(t)
-
-		atx := newChainedActivationTxV1(t, goldenATXID, prevAtx, posAtx.ID())
-		atx.Sign(sig)
-
-		atxHdlr.mclock.EXPECT().CurrentLayer().Return((atx.PublishEpoch - 2).FirstLayer())
-		err := atxHdlr.syntacticallyValidate(context.Background(), atx)
-		require.ErrorContains(t, err, "atx publish epoch is too far in the future")
-	})
-	t.Run("failing nipost challenge validation", func(t *testing.T) {
-		t.Parallel()
-		atxHdlr, prevAtx, posAtx := setup(t)
-
-		atx := newChainedActivationTxV1(t, goldenATXID, prevAtx, posAtx.ID())
-		atx.Sign(sig)
-
-		atxHdlr.mclock.EXPECT().CurrentLayer().Return(atx.PublishEpoch.FirstLayer())
-		require.NoError(t, atxHdlr.syntacticallyValidate(context.Background(), atx))
-
-		atxHdlr.mValidator.EXPECT().
-			NIPostChallengeV1(gomock.Any(), gomock.Any(), atx.SmesherID).
-			Return(errors.New("nipost error"))
-		_, _, proof, err := atxHdlr.syntacticallyValidateDeps(context.Background(), atx)
-		require.EqualError(t, err, "nipost error")
-		require.Nil(t, proof)
-	})
-	t.Run("failing positioning atx validation", func(t *testing.T) {
-		t.Parallel()
-		atxHdlr, prevAtx, posAtx := setup(t)
-
-		atx := newChainedActivationTxV1(t, goldenATXID, prevAtx, posAtx.ID())
-		atx.Sign(sig)
-
-		atxHdlr.mclock.EXPECT().CurrentLayer().Return(atx.PublishEpoch.FirstLayer())
-		require.NoError(t, atxHdlr.syntacticallyValidate(context.Background(), atx))
-
-		atxHdlr.mValidator.EXPECT().NIPostChallengeV1(gomock.Any(), gomock.Any(), atx.SmesherID)
-		atxHdlr.mValidator.EXPECT().
-			PositioningAtx(atx.PositioningATXID, gomock.Any(), goldenATXID, atx.PublishEpoch).
-			Return(errors.New("bad positioning atx"))
-		_, _, proof, err := atxHdlr.syntacticallyValidateDeps(context.Background(), atx)
-		require.EqualError(t, err, "bad positioning atx")
-		require.Nil(t, proof)
-	})
-	t.Run("bad initial nipost challenge", func(t *testing.T) {
-		t.Parallel()
-		atxHdlr, _, posAtx := setup(t)
-
-		cATX := posAtx.ID()
-		atx := newInitialATXv1(t, cATX)
-		atx.Sign(sig)
-
-		atxHdlr.mclock.EXPECT().CurrentLayer().Return(atx.PublishEpoch.FirstLayer())
-		atxHdlr.mValidator.EXPECT().
-			Post(gomock.Any(), sig.NodeID(), cATX, gomock.Any(), gomock.Any(), atx.NumUnits, gomock.Any())
-		atxHdlr.mValidator.EXPECT().VRFNonce(sig.NodeID(), cATX, *atx.VRFNonce, gomock.Any(), atx.NumUnits)
-		require.NoError(t, atxHdlr.syntacticallyValidate(context.Background(), atx))
-
-		atxHdlr.mValidator.EXPECT().
-			InitialNIPostChallengeV1(gomock.Any(), gomock.Any(), goldenATXID).
-			Return(errors.New("bad initial nipost"))
-		_, _, proof, err := atxHdlr.syntacticallyValidateDeps(context.Background(), atx)
-		require.EqualError(t, err, "bad initial nipost")
-		require.Nil(t, proof)
-	})
-	t.Run("bad NIPoST", func(t *testing.T) {
-		t.Parallel()
-		atxHdlr, prevATX, postAtx := setup(t)
-
-		atx := newChainedActivationTxV1(t, goldenATXID, prevATX, postAtx.ID())
-		atx.Sign(sig)
-
-		atxHdlr.mclock.EXPECT().CurrentLayer().Return(atx.PublishEpoch.FirstLayer())
-		require.NoError(t, atxHdlr.syntacticallyValidate(context.Background(), atx))
-
-		atxHdlr.mValidator.EXPECT().NIPostChallengeV1(gomock.Any(), gomock.Any(), atx.SmesherID)
-		atxHdlr.mValidator.EXPECT().PositioningAtx(atx.PositioningATXID, gomock.Any(), goldenATXID, atx.PublishEpoch)
-		atxHdlr.mValidator.EXPECT().
-			NIPost(gomock.Any(), atx.SmesherID, goldenATXID, gomock.Any(), gomock.Any(), atx.NumUnits, gomock.Any()).
-			Return(0, errors.New("bad nipost"))
-		_, _, _, err := atxHdlr.syntacticallyValidateDeps(context.Background(), atx)
-		require.EqualError(t, err, "invalid nipost: bad nipost")
-	})
-	t.Run("can't find VRF nonce", func(t *testing.T) {
-		t.Parallel()
-		atxHdlr, prevATX, postAtx := setup(t)
-
-		atx := newChainedActivationTxV1(t, goldenATXID, prevATX, postAtx.ID())
-		atx.NumUnits += 100
-		atx.Sign(sig)
-
-		enc := func(stmt *sql.Statement) { stmt.BindBytes(1, atx.SmesherID.Bytes()) }
-		_, err := atxHdlr.cdb.Exec(`UPDATE atxs SET nonce = NULL WHERE pubkey = ?1;`, enc, nil)
-		require.NoError(t, err)
-
-		atxHdlr.mclock.EXPECT().CurrentLayer().Return(atx.PublishEpoch.FirstLayer())
-		require.NoError(t, atxHdlr.syntacticallyValidate(context.Background(), atx))
-
-		atxHdlr.mValidator.EXPECT().NIPostChallengeV1(gomock.Any(), gomock.Any(), atx.SmesherID)
-		_, _, _, err1 := atxHdlr.syntacticallyValidateDeps(context.Background(), atx)
-		require.ErrorContains(t, err1, "failed to get current nonce")
-	})
-	t.Run("missing NodeID in initial atx", func(t *testing.T) {
-		t.Parallel()
-		atxHdlr, _, _ := setup(t)
-
-		atx := newInitialATXv1(t, goldenATXID)
-		atx.Signature = sig.Sign(signing.ATX, atx.SignedBytes())
-		atx.SmesherID = sig.NodeID()
-
-		atxHdlr.mclock.EXPECT().CurrentLayer().Return(atx.PublishEpoch.FirstLayer())
-		err := atxHdlr.syntacticallyValidate(context.Background(), atx)
-		require.ErrorContains(t, err, "node id is missing")
-	})
-	t.Run("missing VRF nonce in initial atx", func(t *testing.T) {
-		t.Parallel()
-		atxHdlr, _, _ := setup(t)
-
-		atx := newInitialATXv1(t, goldenATXID)
-		atx.VRFNonce = nil
-		atx.Sign(sig)
-
-		atxHdlr.mclock.EXPECT().CurrentLayer().Return(atx.PublishEpoch.FirstLayer())
-		err := atxHdlr.syntacticallyValidate(context.Background(), atx)
-		require.ErrorContains(t, err, "vrf nonce is missing")
-	})
-	t.Run("invalid VRF nonce in initial atx", func(t *testing.T) {
-		t.Parallel()
-		atxHdlr, _, _ := setup(t)
-
-		atx := newInitialATXv1(t, goldenATXID)
-		atx.Sign(sig)
-
-		atxHdlr.mclock.EXPECT().CurrentLayer().Return(atx.PublishEpoch.FirstLayer())
-		atxHdlr.mValidator.EXPECT().
-			VRFNonce(atx.SmesherID, *atx.CommitmentATXID, *atx.VRFNonce, gomock.Any(), atx.NumUnits).
-			Return(errors.New("invalid VRF nonce"))
-		err := atxHdlr.syntacticallyValidate(context.Background(), atx)
-		require.ErrorContains(t, err, "invalid VRF nonce")
-	})
-	t.Run("prevAtx not declared but initial Post not included", func(t *testing.T) {
-		t.Parallel()
-		atxHdlr, prevAtx, posAtx := setup(t)
-
-		atx := newChainedActivationTxV1(t, goldenATXID, prevAtx, posAtx.ID())
-		atx.PrevATXID = types.EmptyATXID
-		atx.Sign(sig)
-
-		atxHdlr.mclock.EXPECT().CurrentLayer().Return(atx.PublishEpoch.FirstLayer())
-		err := atxHdlr.syntacticallyValidate(context.Background(), atx)
-		require.EqualError(t, err, "no prev atx declared, but initial post is not included")
-	})
-	t.Run("prevAtx not declared but commitment ATX is not included", func(t *testing.T) {
-		t.Parallel()
-		atxHdlr, _, _ := setup(t)
-
-		atx := newInitialATXv1(t, goldenATXID)
-		atx.CommitmentATXID = nil
-		atx.Sign(sig)
-
-		atxHdlr.mclock.EXPECT().CurrentLayer().Return(atx.PublishEpoch.FirstLayer())
-		err := atxHdlr.syntacticallyValidate(context.Background(), atx)
-		require.EqualError(t, err, "no prev atx declared, but commitment atx is missing")
-	})
-	t.Run("prevAtx not declared but commitment ATX is empty", func(t *testing.T) {
-		t.Parallel()
-		atxHdlr, _, _ := setup(t)
-
-		atx := newInitialATXv1(t, goldenATXID)
-		atx.CommitmentATXID = &types.EmptyATXID
-		atx.Sign(sig)
-
-		atxHdlr.mclock.EXPECT().CurrentLayer().Return(atx.PublishEpoch.FirstLayer())
-		err := atxHdlr.syntacticallyValidate(context.Background(), atx)
-		require.EqualError(t, err, "empty commitment atx")
-	})
-	t.Run("prevAtx not declared but sequence not zero", func(t *testing.T) {
-		t.Parallel()
-		atxHdlr, _, _ := setup(t)
-
-		atx := newInitialATXv1(t, goldenATXID)
-		atx.Sequence = 1
-		atx.Sign(sig)
-
-		atxHdlr.mclock.EXPECT().CurrentLayer().Return(atx.PublishEpoch.FirstLayer())
-		err := atxHdlr.syntacticallyValidate(context.Background(), atx)
-		require.EqualError(t, err, "no prev atx declared, but sequence number not zero")
-	})
-	t.Run("prevAtx not declared but validation of initial post fails", func(t *testing.T) {
-		t.Parallel()
-		atxHdlr, _, _ := setup(t)
-
-		atx := newInitialATXv1(t, goldenATXID)
-		atx.Sign(sig)
-
-		atxHdlr.mclock.EXPECT().CurrentLayer().Return(atx.PublishEpoch.FirstLayer())
-		atxHdlr.mValidator.EXPECT().
-			VRFNonce(atx.SmesherID, *atx.CommitmentATXID, *atx.VRFNonce, gomock.Any(), atx.NumUnits)
-		atxHdlr.mValidator.EXPECT().
-			Post(gomock.Any(), atx.SmesherID, gomock.Any(), gomock.Any(), gomock.Any(), atx.NumUnits, gomock.Any()).
-			Return(errors.New("failed post validation"))
-		err := atxHdlr.syntacticallyValidate(context.Background(), atx)
-		require.ErrorContains(t, err, "failed post validation")
-	})
-	t.Run("empty positioning ATX", func(t *testing.T) {
-		t.Parallel()
-		atxHdlr, _, _ := setup(t)
-
-		atx := newInitialATXv1(t, goldenATXID)
-		atx.PositioningATXID = types.EmptyATXID
-		atx.Sign(sig)
-
-		atxHdlr.mclock.EXPECT().CurrentLayer().Return(atx.PublishEpoch.FirstLayer())
-		err := atxHdlr.syntacticallyValidate(context.Background(), atx)
-		require.EqualError(t, err, "empty positioning atx")
-	})
-	t.Run("prevAtx declared but initial Post is included", func(t *testing.T) {
-		t.Parallel()
-		atxHdlr, prevAtx, _ := setup(t)
-
-		atx := newInitialATXv1(t, goldenATXID)
-		atx.PrevATXID = prevAtx.ID()
-		atx.Sign(sig)
-
-		atxHdlr.mclock.EXPECT().CurrentLayer().Return(atx.PublishEpoch.FirstLayer())
-		err := atxHdlr.syntacticallyValidate(context.Background(), atx)
-		require.EqualError(t, err, "prev atx declared, but initial post is included")
-	})
-	t.Run("prevAtx declared but NodeID is included", func(t *testing.T) {
-		t.Parallel()
-		atxHdlr, prevAtx, posAtx := setup(t)
-
-		atx := newChainedActivationTxV1(t, goldenATXID, prevAtx, posAtx.ID())
-		atx.NodeID = &types.NodeID{1, 2, 3}
-		atx.Sign(sig)
-
-		atxHdlr.mclock.EXPECT().CurrentLayer().Return(atx.PublishEpoch.FirstLayer())
-		err := atxHdlr.syntacticallyValidate(context.Background(), atx)
-		require.EqualError(t, err, "prev atx declared, but node id is included")
-	})
-	t.Run("prevAtx declared but commitmentATX is included", func(t *testing.T) {
-		t.Parallel()
-		atxHdlr, prevAtx, posAtx := setup(t)
-
-		atx := newChainedActivationTxV1(t, goldenATXID, prevAtx, posAtx.ID())
-		atx.CommitmentATXID = &types.EmptyATXID
-		atx.Sign(sig)
-
-		atxHdlr.mclock.EXPECT().CurrentLayer().Return(atx.PublishEpoch.FirstLayer())
-		err := atxHdlr.syntacticallyValidate(context.Background(), atx)
-		require.EqualError(t, err, "prev atx declared, but commitment atx is included")
-	})
-}
-
-func TestHandler_ContextuallyValidateAtx(t *testing.T) {
-	goldenATXID := types.ATXID{2, 3, 4}
-
-	sig, err := signing.NewEdSigner()
-	require.NoError(t, err)
-
-	t.Run("valid initial atx", func(t *testing.T) {
-		t.Parallel()
-
-		atx := newInitialATXv1(t, goldenATXID)
-		atx.Sign(sig)
-
-		atxHdlr := newTestHandler(t, goldenATXID)
-		require.NoError(t, atxHdlr.contextuallyValidateAtx(atx))
-	})
-
-	t.Run("missing prevAtx", func(t *testing.T) {
-		t.Parallel()
-
-		atxHdlr := newTestHandler(t, goldenATXID)
-
-		prevAtx := newInitialATXv1(t, goldenATXID)
-		atx := newChainedActivationTxV1(t, goldenATXID, prevAtx, goldenATXID)
-
-		err = atxHdlr.contextuallyValidateAtx(atx)
-		require.ErrorIs(t, err, sql.ErrNotFound)
-	})
-
-	t.Run("wrong previous atx by same node", func(t *testing.T) {
-		t.Parallel()
-
-		atxHdlr := newTestHandler(t, goldenATXID)
-
-		atx0 := newInitialATXv1(t, goldenATXID)
-		atx0.Sign(sig)
-		atxHdlr.expectAtxV1(atx0, sig.NodeID())
-		require.NoError(t, atxHdlr.HandleGossipAtx(context.Background(), "", codec.MustEncode(atx0)))
-
-		atx1 := newChainedActivationTxV1(t, goldenATXID, atx0, goldenATXID)
-		atx1.Sign(sig)
-		atxHdlr.expectAtxV1(atx1, sig.NodeID())
-		atxHdlr.mockFetch.EXPECT().GetAtxs(gomock.Any(), gomock.Any(), gomock.Any())
-		require.NoError(t, atxHdlr.HandleGossipAtx(context.Background(), "", codec.MustEncode(atx1)))
-
-		atxInvalidPrevious := newChainedActivationTxV1(t, goldenATXID, atx0, goldenATXID)
-		atxInvalidPrevious.Sign(sig)
-		err := atxHdlr.contextuallyValidateAtx(atxInvalidPrevious)
-		require.EqualError(t, err, "last atx is not the one referenced")
-	})
-
-	t.Run("wrong previous atx from different node", func(t *testing.T) {
-		t.Parallel()
-
-		otherSig, err := signing.NewEdSigner()
-		require.NoError(t, err)
-
-		atxHdlr := newTestHandler(t, goldenATXID)
-
-		atx0 := newInitialATXv1(t, goldenATXID)
-		atx0.Sign(otherSig)
-		atxHdlr.expectAtxV1(atx0, otherSig.NodeID())
-		require.NoError(t, atxHdlr.HandleGossipAtx(context.Background(), "", codec.MustEncode(atx0)))
-
-		atx1 := newInitialATXv1(t, goldenATXID)
-		atx1.Sign(sig)
-		atxHdlr.expectAtxV1(atx1, sig.NodeID())
-		require.NoError(t, atxHdlr.HandleGossipAtx(context.Background(), "", codec.MustEncode(atx1)))
-
-		atxInvalidPrevious := newChainedActivationTxV1(t, goldenATXID, atx0, goldenATXID)
-		atxInvalidPrevious.Sign(sig)
-		err = atxHdlr.contextuallyValidateAtx(atxInvalidPrevious)
-		require.EqualError(t, err, "last atx is not the one referenced")
-	})
-}
-
-func TestHandler_StoreAtx(t *testing.T) {
-	goldenATXID := types.RandomATXID()
-
-	sig, err := signing.NewEdSigner()
-	require.NoError(t, err)
-
-	t.Run("stores ATX in DB", func(t *testing.T) {
-		atxHdlr := newTestHandler(t, goldenATXID)
-
-		watx := newInitialATXv1(t, goldenATXID)
-		watx.Sign(sig)
-		vAtx := toAtx(t, watx)
-		require.NoError(t, err)
-
-		atxHdlr.mbeacon.EXPECT().OnAtx(vAtx.ToHeader())
-		atxHdlr.mtortoise.EXPECT().OnAtx(gomock.Any(), vAtx.ID(), gomock.Any())
-		proof, err := atxHdlr.storeAtx(context.Background(), vAtx)
-		require.NoError(t, err)
-		require.Nil(t, proof)
-
-		atxFromDb, err := atxs.Get(atxHdlr.cdb, vAtx.ID())
-		require.NoError(t, err)
-		vAtx.SetReceived(time.Unix(0, vAtx.Received().UnixNano()))
-		require.Equal(t, vAtx, atxFromDb)
-	})
-
-	t.Run("storing an already known ATX returns no error", func(t *testing.T) {
-		atxHdlr := newTestHandler(t, goldenATXID)
-
-		watx := newInitialATXv1(t, goldenATXID)
-		watx.Sign(sig)
-		vAtx := toAtx(t, watx)
-
-		atxHdlr.mbeacon.EXPECT().OnAtx(vAtx.ToHeader())
-		atxHdlr.mtortoise.EXPECT().OnAtx(gomock.Any(), vAtx.ID(), gomock.Any())
-		proof, err := atxHdlr.storeAtx(context.Background(), vAtx)
-		require.NoError(t, err)
-		require.Nil(t, proof)
-
-		atxHdlr.mbeacon.EXPECT().OnAtx(vAtx.ToHeader())
-		// Note: tortoise is not informed about the same ATX again
-		proof, err = atxHdlr.storeAtx(context.Background(), vAtx)
-		require.NoError(t, err)
-		require.Nil(t, proof)
-	})
-
-	t.Run("another atx for the same epoch is considered malicious", func(t *testing.T) {
-		atxHdlr := newTestHandler(t, goldenATXID)
-
-		watx0 := newInitialATXv1(t, goldenATXID)
-		watx0.Sign(sig)
-		vAtx0 := toAtx(t, watx0)
-
-		atxHdlr.mbeacon.EXPECT().OnAtx(vAtx0.ToHeader())
-		atxHdlr.mtortoise.EXPECT().OnAtx(gomock.Any(), vAtx0.ID(), gomock.Any())
-
-		proof, err := atxHdlr.storeAtx(context.Background(), vAtx0)
-		require.NoError(t, err)
-		require.Nil(t, proof)
-
-		watx1 := newInitialATXv1(t, goldenATXID)
-		watx1.Coinbase = types.GenerateAddress([]byte("aaaa"))
-		watx1.Sign(sig)
-		vAtx1 := toAtx(t, watx1)
-
-		atxHdlr.mbeacon.EXPECT().OnAtx(vAtx1.ToHeader())
-		atxHdlr.mtortoise.EXPECT().OnAtx(gomock.Any(), vAtx1.ID(), gomock.Any())
-		atxHdlr.mtortoise.EXPECT().OnMalfeasance(sig.NodeID())
-
-		proof, err = atxHdlr.storeAtx(context.Background(), vAtx1)
-		require.NoError(t, err)
-		require.NotNil(t, proof)
-		require.Equal(t, mwire.MultipleATXs, proof.Proof.Type)
-
-		proof.SetReceived(time.Time{})
-		nodeID, err := malfeasance.Validate(
-			context.Background(),
-			atxHdlr.log,
-			atxHdlr.cdb,
-			atxHdlr.edVerifier,
-			nil,
-			&mwire.MalfeasanceGossip{MalfeasanceProof: *proof},
-		)
-		require.NoError(t, err)
-		require.Equal(t, sig.NodeID(), nodeID)
-
-		malicious, err := identities.IsMalicious(atxHdlr.cdb, sig.NodeID())
-		require.NoError(t, err)
-		require.True(t, malicious)
-	})
-
-	t.Run("another atx for the same epoch for registered ID doesn't create a malfeasance proof", func(t *testing.T) {
-		atxHdlr := newTestHandler(t, goldenATXID)
-		atxHdlr.Register(sig)
-
-		watx0 := newInitialATXv1(t, goldenATXID)
-		watx0.Sign(sig)
-		vAtx0 := toAtx(t, watx0)
-
-		atxHdlr.mbeacon.EXPECT().OnAtx(vAtx0.ToHeader())
-		atxHdlr.mtortoise.EXPECT().OnAtx(gomock.Any(), vAtx0.ID(), gomock.Any())
-
-		proof, err := atxHdlr.storeAtx(context.Background(), vAtx0)
-		require.NoError(t, err)
-		require.Nil(t, proof)
-
-		watx1 := newInitialATXv1(t, goldenATXID)
-		watx1.Coinbase = types.GenerateAddress([]byte("aaaa"))
-		watx1.Sign(sig)
-		vAtx1 := toAtx(t, watx1)
-
-		proof, err = atxHdlr.storeAtx(context.Background(), vAtx1)
-		require.ErrorContains(t,
-			err,
-			fmt.Sprintf("%s already published an ATX", sig.NodeID().ShortString()),
-		)
-		require.Nil(t, proof)
-
-		malicious, err := identities.IsMalicious(atxHdlr.cdb, sig.NodeID())
-		require.NoError(t, err)
-		require.False(t, malicious)
-	})
 }
 
 func testHandler_PostMalfeasanceProofs(t *testing.T, synced bool) {
@@ -1209,71 +618,6 @@ func TestCollectDeps(t *testing.T) {
 	})
 }
 
-func TestHandler_RegistersHashesInPeer(t *testing.T) {
-	goldenATXID := types.RandomATXID()
-	peer := p2p.Peer("buddy")
-	t.Run("registers poet and atxs", func(t *testing.T) {
-		t.Parallel()
-		atxHdlr := newTestHandler(t, goldenATXID)
-
-		poet := types.RandomHash()
-		atxs := []types.ATXID{types.RandomATXID(), types.RandomATXID()}
-
-		atxHdlr.mockFetch.EXPECT().
-			RegisterPeerHashes(peer, gomock.InAnyOrder([]types.Hash32{poet, atxs[0].Hash32(), atxs[1].Hash32()}))
-		atxHdlr.registerHashes(peer, poet, atxs)
-	})
-	t.Run("registers poet", func(t *testing.T) {
-		t.Parallel()
-		atxHdlr := newTestHandler(t, goldenATXID)
-
-		poet := types.RandomHash()
-
-		atxHdlr.mockFetch.EXPECT().RegisterPeerHashes(peer, []types.Hash32{poet})
-		atxHdlr.registerHashes(peer, poet, nil)
-	})
-}
-
-func TestHandler_FetchesReferences(t *testing.T) {
-	goldenATXID := types.RandomATXID()
-	t.Run("fetch poet and atxs", func(t *testing.T) {
-		t.Parallel()
-		atxHdlr := newTestHandler(t, goldenATXID)
-
-		poet := types.RandomHash()
-		atxs := []types.ATXID{types.RandomATXID(), types.RandomATXID()}
-
-		atxHdlr.mockFetch.EXPECT().GetPoetProof(gomock.Any(), poet)
-		atxHdlr.mockFetch.EXPECT().GetAtxs(gomock.Any(), atxs, gomock.Any())
-		require.NoError(t, atxHdlr.fetchReferences(context.Background(), poet, atxs))
-	})
-
-	t.Run("no poet proofs", func(t *testing.T) {
-		t.Parallel()
-		atxHdlr := newTestHandler(t, goldenATXID)
-
-		poet := types.RandomHash()
-
-		atxHdlr.mockFetch.EXPECT().GetPoetProof(gomock.Any(), poet).Return(errors.New("pooh"))
-		require.Error(t, atxHdlr.fetchReferences(context.Background(), poet, nil))
-	})
-
-	t.Run("no atxs", func(t *testing.T) {
-		t.Parallel()
-		atxHdlr := newTestHandler(t, goldenATXID)
-
-		poet := types.RandomHash()
-		atxs := []types.ATXID{types.RandomATXID(), types.RandomATXID()}
-
-		atxHdlr.mockFetch.EXPECT().GetPoetProof(gomock.Any(), poet).Return(errors.New("pooh"))
-		require.Error(t, atxHdlr.fetchReferences(context.Background(), poet, nil))
-
-		atxHdlr.mockFetch.EXPECT().GetPoetProof(gomock.Any(), poet)
-		atxHdlr.mockFetch.EXPECT().GetAtxs(gomock.Any(), atxs, gomock.Any()).Return(errors.New("oh"))
-		require.Error(t, atxHdlr.fetchReferences(context.Background(), poet, atxs))
-	})
-}
-
 func TestHandler_AtxWeight(t *testing.T) {
 	const (
 		peer     = p2p.Peer("buddy")
@@ -1283,8 +627,7 @@ func TestHandler_AtxWeight(t *testing.T) {
 	)
 
 	goldenATXID := types.ATXID{2, 3, 4}
-	atxHdlr := newTestHandler(t, goldenATXID)
-	atxHdlr.tickSize = tickSize
+	atxHdlr := newTestHandler(t, goldenATXID, func(o *testHandlerOptions) { o.tickSize = tickSize })
 
 	sig, err := signing.NewEdSigner()
 	require.NoError(t, err)
